+++ src/aims_ui/page_controllers/b_multiple_matches/utils/multiple_match_api_utils.py
@@ -3,7 +3,7 @@ import json
 import logging
 
 from aims_ui.page_helpers.api.api_helpers import job_api
-from aims_ui.app_helpers.google_utils import get_username
+from aims_ui.page_helpers.google_utils import get_username
 from aims_ui.page_helpers.api.api_helpers import get_header
